requirements.txt
eventlet~=0.40.3
flask-socketio==5.3.6
Flask==3.0.0
python-socketio[client]
requests
